Arif plaza Inn web intelligence
The website had to be competitive and this needed use of AI for
understanding customer interactions, booking patterns and based on the
dataset collected.
Use of ML algorithms to classify, normalize, display and provide
customers a better user experience by presenting to them features/
services /menus as per the recommended results of these algorithms;
leveraged NoSQL for storing of unstructured data, retrieval and display,
ML algorithms for design of recommender engine and predicting the
outcomes. Use of Python recommender based on bookings data – Python
libraries like panda, JavaScript lib. like LAMP/ React used to design
and develop the front and back end